requirements.txt
discord.py==1.7.3
PyNaCl==1.4.0
ffmpeg-python==0.2.0